# artifacts/api-server/src/routes/gifts.ts
import { Router } from "express";
import { db, giftsTable, giftItemsTable, usersTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";
import { SendGiftBody } from "@workspace/api-zod";

const router = Router();
const CURRENT_USER_ID = 1;

async function buildGift(gift: typeof giftsTable.$inferSelect) {
  const giftItem = await db.query.giftItemsTable.findFirst({ where: eq(giftItemsTable.id, gift.giftItemId) });
  const sender = gift.isAnonymous ? null : await db.query.usersTable.findFirst({ where: eq(usersTable.id, gift.senderId) });
  const receiver = await db.query.usersTable.findFirst({ where: eq(usersTable.id, gift.receiverId) });
  return { ...gift, giftItem, sender, receiver };
}

router.get("/gifts", async (req, res) => {
  try {
    const items = await db.select().from(giftItemsTable);
    res.json(items);
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/gifts/sent", async (req, res) => {
  try {
    const gifts = await db.select().from(giftsTable)
      .where(eq(giftsTable.senderId, CURRENT_USER_ID))
      .orderBy(desc(giftsTable.createdAt));
    const built = await Promise.all(gifts.map(buildGift));
    res.json(built);
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/gifts/received", async (req, res) => {
  try {
    const gifts = await db.select().from(giftsTable)
      .where(eq(giftsTable.receiverId, CURRENT_USER_ID))
      .orderBy(desc(giftsTable.createdAt));
    const built = await Promise.all(gifts.map(buildGift));
    res.json(built);
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/gifts/send", async (req, res) => {
  try {
    const body = SendGiftBody.parse(req.body);
    const [gift] = await db.insert(giftsTable).values({
      giftItemId: body.giftItemId,
      senderId: CURRENT_USER_ID,
      receiverId: body.receiverId,
      message: body.message,
      isAnonymous: body.isAnonymous ?? false,
      chatId: body.chatId,
    }).returning();
    const built = await buildGift(gift);
    res.status(201).json(built);
  } catch (err) {
    req.log.error(err);
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;